back-end/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.3
tzdata==2022.7
